Refactored policies for easier testing and maintainability
By having objects per-policy, it makes it much easier to maintain each
policy. In effect, each policy votes on whether to keep or delete a
filename. If all policies vote to delete a file, then it will be
reported as deleteable.

This change means we don't do stream processing anymore: the total
result set is read in memory all at once and processed then reported.
The system does not do stream processing anymore.